Adding chatroom initial code
