+++ example/pubspec_overrides.yaml
@@ -2,8 +2,10 @@
 
 dependency_overrides:
   extended_text_field: ^15.0.0
-  js: ^0.7.1
   likeminds_chat_flutter_core:
     path: ../lib/packages/core
   likeminds_chat_flutter_ui:
     path: ../lib/packages/ui
+  likeminds_chat_fl:
+    path: ../../LikeMinds-Flutter-GroupChat-SDK/
+  js: ^0.7.1
