+++ lib/packages/core/.flutter-plugins
@@ -1,41 +1,40 @@
 # This is a generated file; do not edit or check into version control.
-connectivity_plus=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/connectivity_plus-6.0.3/
 device_info_plus=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/device_info_plus-10.1.0/
 file_picker=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/file_picker-8.0.3/
 file_selector_linux=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/file_selector_linux-0.9.2+1/
 file_selector_macos=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/file_selector_macos-0.9.4/
 file_selector_windows=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/file_selector_windows-0.9.3+1/
-firebase_core=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_core-2.31.0/
-firebase_core_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_core_web-2.17.0/
-firebase_database=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_database-10.5.5/
-firebase_database_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_database_web-0.2.5+5/
-firebase_messaging=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_messaging-14.9.2/
-firebase_messaging_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_messaging_web-3.8.5/
+firebase_core=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_core-2.32.0/
+firebase_core_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_core_web-2.17.1/
+firebase_database=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_database-10.5.7/
+firebase_database_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_database_web-0.2.5+7/
+firebase_messaging=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_messaging-14.9.4/
+firebase_messaging_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/firebase_messaging_web-3.8.7/
 flutter_keyboard_visibility=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/flutter_keyboard_visibility-6.0.0/
 flutter_keyboard_visibility_linux=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/flutter_keyboard_visibility_linux-1.0.0/
 flutter_keyboard_visibility_macos=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/flutter_keyboard_visibility_macos-1.0.0/
 flutter_keyboard_visibility_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/flutter_keyboard_visibility_web-2.0.0/
 flutter_keyboard_visibility_windows=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/flutter_keyboard_visibility_windows-1.0.0/
-flutter_plugin_android_lifecycle=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.19/
-image_picker=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker-1.1.1/
-image_picker_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker_android-0.8.12+1/
+flutter_plugin_android_lifecycle=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.20/
+image_picker=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker-1.1.2/
+image_picker_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker_android-0.8.12+3/
 image_picker_for_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker_for_web-3.0.4/
-image_picker_ios=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker_ios-0.8.11+2/
+image_picker_ios=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker_ios-0.8.12/
 image_picker_linux=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker_linux-0.2.1+1/
 image_picker_macos=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker_macos-0.2.1+1/
 image_picker_windows=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/image_picker_windows-0.2.1+1/
 media_kit_video=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/media_kit_video-1.2.4/
 package_info_plus=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/package_info_plus-8.0.0/
 path_provider=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/path_provider-2.1.3/
-path_provider_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/path_provider_android-2.2.4/
+path_provider_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/path_provider_android-2.2.5/
 path_provider_foundation=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/path_provider_foundation-2.4.0/
 path_provider_linux=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/path_provider_linux-2.2.1/
 path_provider_windows=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/path_provider_windows-2.2.1/
 pdf_bitmaps=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/pdf_bitmaps-1.0.1/
-pdf_render=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/pdf_render-1.4.7/
+pdf_render=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/pdf_render-1.4.8/
 permission_handler=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/permission_handler-11.3.1/
 permission_handler_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/permission_handler_android-12.0.6/
-permission_handler_apple=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/permission_handler_apple-9.4.4/
+permission_handler_apple=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/permission_handler_apple-9.4.5/
 permission_handler_html=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/permission_handler_html-0.1.1/
 permission_handler_windows=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/permission_handler_windows-0.2.1/
 screen_brightness=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/screen_brightness-0.2.2+1/
@@ -44,15 +43,15 @@ screen_brightness_ios=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/screen_bri
 screen_brightness_macos=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/screen_brightness_macos-0.1.0+1/
 screen_brightness_windows=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/screen_brightness_windows-0.1.3/
 shared_preferences=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/shared_preferences-2.2.3/
-shared_preferences_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/shared_preferences_android-2.2.2/
+shared_preferences_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/shared_preferences_android-2.2.3/
 shared_preferences_foundation=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/shared_preferences_foundation-2.4.0/
 shared_preferences_linux=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/shared_preferences_linux-2.3.2/
 shared_preferences_web=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/shared_preferences_web-2.3.0/
 shared_preferences_windows=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/shared_preferences_windows-2.3.2/
 simple_s3=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/simple_s3-0.3.2/
 sqflite=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/sqflite-2.3.3+1/
-url_launcher=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/url_launcher-6.2.6/
-url_launcher_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/url_launcher_android-6.3.1/
+url_launcher=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/url_launcher-6.3.0/
+url_launcher_android=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/url_launcher_android-6.3.3/
 url_launcher_ios=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/url_launcher_ios-6.3.0/
 url_launcher_linux=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/url_launcher_linux-3.1.1/
 url_launcher_macos=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/url_launcher_macos-3.2.0/
