+++ lib/packages/core/lib/src/blocs/chatroom/chatroom_bloc.dart
@@ -2,24 +2,32 @@ import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 import 'package:likeminds_chat_fl/likeminds_chat_fl.dart';
+import 'package:likeminds_chat_flutter_core/src/convertors/convertors.dart';
 import 'package:likeminds_chat_flutter_core/src/core/core.dart';
+import 'package:likeminds_chat_flutter_core/src/utils/utils.dart';
+import 'package:likeminds_chat_flutter_ui/likeminds_chat_flutter_ui.dart';
 
 part 'chatroom_event.dart';
 part 'chatroom_state.dart';
 
+part 'handler/fetch_chatroom_handler.dart';
+
+/// BLoC responsible for handling the Chatroom,
+/// Allows for users to fetch and display a chatroom,
 class LMChatroomBloc extends Bloc<LMChatroomEvent, LMChatroomState> {
   static LMChatroomBloc? _instance;
-  static LMChatroomBloc get instance => _instance ??= LMChatroomBloc._();
-  LMChatroomBloc._() : super(LMChatroomInitialState()) {
-    on<LMChatroomEvent>((event, emit) async {
-      if (event is LMChatInitChatroomEvent) {
-        emit(LMChatroomLoadingState());
-        LMResponse<GetChatroomResponse> getChatroomResponse =
-            await LMChatCore.client.getChatroom(event.chatroomRequest);
-        emit(LMChatroomLoadedState(
-          getChatroomResponse: getChatroomResponse.data!,
-        ));
-      }
-    });
+
+  // Creating a singleton instance of the LMChatHomeFeedBloc
+  static LMChatroomBloc get instance {
+    if (_instance == null || _instance!.isClosed) {
+      return _instance = LMChatroomBloc._();
+    } else {
+      return _instance!;
+    }
+  }
+
+  LMChatroomBloc._() : super(LMChatroomLoadingState()) {
+    // Event handler for fetching chatroom
+    on<LMChatFetchChatroomEvent>(fetchChatroomEventHandler);
   }
 }
