+++ lib/packages/core/lib/src/blocs/conversation/conversation_bloc.dart
@@ -14,8 +14,11 @@ import 'package:likeminds_chat_flutter_core/src/utils/utils.dart';
 part 'conversation_event.dart';
 part 'conversation_state.dart';
 
+part 'handler/post_conversation_handler.dart';
+part 'handler/post_multimedia_conversation_handler.dart';
+
 class LMChatConversationBloc
-    extends Bloc<ConversationEvent, LMChatConversationState> {
+    extends Bloc<LMChatConversationEvent, LMChatConversationState> {
   final mediaService = LMChatAWSUtility(!isDebug);
   final DatabaseReference realTime = LMChatRealtime.instance.chatroom();
   int? lastConversationId;
@@ -23,31 +26,31 @@ class LMChatConversationBloc
   static LMChatConversationBloc get instance =>
       _instance ??= LMChatConversationBloc._();
   LMChatConversationBloc._() : super(ConversationInitial()) {
-    on<InitConversations>(
-      (event, emit) {
-        debugPrint("Conversations initiated");
-        int chatroomId = event.chatroomId;
-        lastConversationId = event.conversationId;
+    // on<InitConversations>(
+    // (event, emit) {
+    // debugPrint("Conversations initiated");
+    // int chatroomId = event.chatroomId;
+    // lastConversationId = event.conversationId;
 
-        realTime.onValue.listen(
-          (event) {
-            if (event.snapshot.value != null) {
-              final response = event.snapshot.value as Map;
-              final conversationId =
-                  int.parse(response["collabcard"]["answer_id"]);
-              if (lastConversationId != null &&
-                  conversationId != lastConversationId) {
-                add(UpdateConversations(
-                  chatroomId: chatroomId,
-                  conversationId: conversationId,
-                ));
-              }
-            }
-          },
-        );
-      },
-    );
-    on<ConversationEvent>((event, emit) async {
+    //   realTime.onValue.listen(
+    //     (event) {
+    //       if (event.snapshot.value != null) {
+    //         final response = event.snapshot.value as Map;
+    //         final conversationId =
+    //             int.parse(response["collabcard"]["answer_id"]);
+    //         if (lastConversationId != null &&
+    //             conversationId != lastConversationId) {
+    //           add(UpdateConversations(
+    //             chatroomId: chatroomId,
+    //             conversationId: conversationId,
+    //           ));
+    //         }
+    //       }
+    //     },
+    //   );
+    // },
+    // );
+    on<LMChatConversationEvent>((event, emit) async {
       if (event is LoadConversations) {
         if (event.getConversationRequest.page > 1) {
           emit(ConversationPaginationLoading());
@@ -73,13 +76,9 @@ class LMChatConversationBloc
                     element.replyConversationObject?.userId ??
                         element.replyConversationObject?.memberId];
           }
-          emit(
-            ConversationLoaded(conversationResponse),
-          );
+          emit(ConversationLoaded(conversationResponse));
         } else {
-          emit(
-            ConversationError(response.errorMessage!, ''),
-          );
+          emit(ConversationError(response.errorMessage!, ''));
         }
       }
     });
@@ -150,189 +149,5 @@ class LMChatConversationBloc
         }
       },
     );
-  }
-
-  mapPostMultiMediaConversation(
-    PostMultiMediaConversation event,
-    Emitter<LMChatConversationState> emit,
-  ) async {
-    try {
-      DateTime dateTime = DateTime.now();
-      User user = LMChatPreferences.instance.getUser()!;
-      Conversation conversation = Conversation(
-        answer: event.postConversationRequest.text,
-        chatroomId: event.postConversationRequest.chatroomId,
-        createdAt: "",
-        header: "",
-        date: "${dateTime.day} ${dateTime.month} ${dateTime.year}",
-        replyId: event.postConversationRequest.replyId,
-        attachmentCount: event.postConversationRequest.attachmentCount,
-        hasFiles: event.postConversationRequest.hasFiles,
-        member: user,
-        temporaryId: event.postConversationRequest.temporaryId,
-        id: 1,
-        userId: user.id,
-        ogTags: event.postConversationRequest.ogTags,
-      );
-
-      emit(
-        MultiMediaConversationLoading(
-          conversation,
-          event.mediaFiles,
-        ),
-      );
-      LMResponse<PostConversationResponse> response =
-          await LMChatCore.client.postConversation(
-        event.postConversationRequest,
-      );
-
-      if (response.success) {
-        PostConversationResponse postConversationResponse = response.data!;
-        if (event.mediaFiles.length == 1 &&
-            event.mediaFiles.first.mediaType == LMChatMediaType.link) {
-          emit(
-            MultiMediaConversationPosted(
-              postConversationResponse,
-              event.mediaFiles,
-            ),
-          );
-        } else {
-          List<LMChatMedia> fileLink = [];
-          int length = event.mediaFiles.length;
-          for (int i = 0; i < length; i++) {
-            LMChatMedia media = event.mediaFiles[i];
-            String? url = await mediaService.uploadFile(
-              media.mediaFile!,
-              event.postConversationRequest.chatroomId,
-              postConversationResponse.conversation!.id,
-            );
-            String? thumbnailUrl;
-            if (media.mediaType == LMChatMediaType.video) {
-              // If the thumbnail file is not present in media object
-              // then generate the thumbnail and upload it to the server
-              if (media.thumbnailFile == null) {
-                await getVideoThumbnail(media);
-              }
-              thumbnailUrl = await mediaService.uploadFile(
-                media.thumbnailFile!,
-                event.postConversationRequest.chatroomId,
-                postConversationResponse.conversation!.id,
-              );
-            }
-
-            String attachmentType = mapMediaTypeToString(media.mediaType);
-            PutMediaRequest putMediaRequest = (PutMediaRequestBuilder()
-                  ..conversationId(postConversationResponse.conversation!.id)
-                  ..filesCount(length)
-                  ..index(i)
-                  ..height(media.height)
-                  ..width(media.width)
-                  ..meta({
-                    'size': media.size,
-                    'number_of_page': media.pageCount,
-                  })
-                  ..type(attachmentType)
-                  ..thumbnailUrl(thumbnailUrl)
-                  ..url(url!))
-                .build();
-            LMResponse<PutMediaResponse> uploadFileResponse =
-                await LMChatCore.client.putMultimedia(putMediaRequest);
-            if (!uploadFileResponse.success) {
-              emit(
-                MultiMediaConversationError(
-                  uploadFileResponse.errorMessage!,
-                  event.postConversationRequest.temporaryId,
-                ),
-              );
-            } else {
-              emit(
-                MultiMediaConversationError(
-                  uploadFileResponse.errorMessage!,
-                  event.postConversationRequest.temporaryId,
-                ),
-              );
-            }
-          }
-          lastConversationId = response.data!.conversation!.id;
-          emit(
-            MultiMediaConversationPosted(
-              postConversationResponse,
-              fileLink,
-            ),
-          );
-        }
-      } else {
-        emit(
-          MultiMediaConversationError(
-            response.errorMessage!,
-            event.postConversationRequest.temporaryId,
-          ),
-        );
-        return false;
-      }
-    } catch (e) {
-      emit(
-        ConversationError(
-          "An error occurred",
-          event.postConversationRequest.temporaryId,
-        ),
-      );
-      return false;
-    }
-  }
-
-  mapPostConversationFunction(
-      PostConversation event, Emitter<LMChatConversationState> emit) async {
-    try {
-      DateTime dateTime = DateTime.now();
-      User user = LMChatPreferences.instance.getUser()!;
-      Conversation conversation = Conversation(
-        answer: event.postConversationRequest.text,
-        chatroomId: event.postConversationRequest.chatroomId,
-        createdAt: "",
-        userId: user.id,
-        header: "",
-        date: "${dateTime.day} ${dateTime.month} ${dateTime.year}",
-        replyId: event.postConversationRequest.replyId,
-        attachmentCount: event.postConversationRequest.attachmentCount,
-        replyConversationObject: event.repliedTo,
-        hasFiles: event.postConversationRequest.hasFiles,
-        member: user,
-        temporaryId: event.postConversationRequest.temporaryId,
-        id: 1,
-      );
-      emit(LocalConversation(conversation));
-      LMResponse<PostConversationResponse> response =
-          await LMChatCore.client.postConversation(
-        event.postConversationRequest,
-      );
-
-      if (response.success) {
-        Conversation conversation = response.data!.conversation!;
-        if (conversation.replyId != null ||
-            conversation.replyConversation != null) {
-          conversation.replyConversationObject = event.repliedTo;
-        }
-        emit(ConversationPosted(response.data!));
-
-        return false;
-      } else {
-        emit(
-          ConversationError(
-            response.errorMessage!,
-            event.postConversationRequest.temporaryId,
-          ),
-        );
-        return false;
-      }
-    } catch (e) {
-      emit(
-        ConversationError(
-          "An error occurred",
-          event.postConversationRequest.temporaryId,
-        ),
-      );
-      return false;
-    }
   }
 }
