+++ lib/packages/core/lib/src/views/chatroom/chatroom.dart
@@ -7,16 +7,11 @@ import 'package:likeminds_chat_fl/likeminds_chat_fl.dart';
 import 'package:likeminds_chat_flutter_core/src/blocs/blocs.dart';
 import 'package:likeminds_chat_flutter_core/src/blocs/observer.dart';
 import 'package:likeminds_chat_flutter_core/src/convertors/chatroom/chatroom_convertor.dart';
-import 'package:likeminds_chat_flutter_core/src/convertors/conversation/conversation_convertor.dart';
-import 'package:likeminds_chat_flutter_core/src/convertors/user/user_convertor.dart';
-import 'package:likeminds_chat_flutter_core/src/utils/analytics/analytics.dart';
-import 'package:likeminds_chat_flutter_core/src/utils/conversation/conversation_utils.dart';
-import 'package:likeminds_chat_flutter_core/src/utils/preferences/preferences.dart';
 import 'package:likeminds_chat_flutter_core/src/utils/utils.dart';
 import 'package:likeminds_chat_flutter_core/src/widgets/chatroom/chatroom_bar.dart';
 import 'package:likeminds_chat_flutter_core/src/widgets/chatroom/chatroom_menu.dart';
+import 'package:likeminds_chat_flutter_core/src/widgets/lists/conversation/conversation_list.dart';
 import 'package:likeminds_chat_flutter_ui/likeminds_chat_flutter_ui.dart';
-import 'package:overlay_support/overlay_support.dart';
 
 class LMChatroomScreen extends StatefulWidget {
   final int chatroomId;
@@ -46,8 +41,6 @@ class LMChatroomScreen extends StatefulWidget {
 }
 
 class _LMChatroomScreenState extends State<LMChatroomScreen> {
-  late LMChatConversationBloc _conversationBloc;
-  late LMChatConversationActionBloc _convActionBloc;
   late LMChatroomBloc _chatroomBloc;
   late LMChatroomActionBloc _chatroomActionBloc;
 
@@ -76,20 +69,13 @@ class _LMChatroomScreenState extends State<LMChatroomScreen> {
   PagingController<int, Conversation> pagedListController =
       PagingController<int, Conversation>(firstPageKey: 1);
 
-  int _page = 1;
-  ModalRoute? _route;
-
   @override
   void initState() {
     super.initState();
     Bloc.observer = LMChatBlocObserver();
-    _conversationBloc = LMChatConversationBloc.instance;
     _chatroomBloc = LMChatroomBloc.instance;
-    _convActionBloc = LMChatConversationActionBloc.instance;
     _chatroomActionBloc = LMChatroomActionBloc.instance;
-    _chatroomBloc.add(LMChatInitChatroomEvent(
-        (GetChatroomRequestBuilder()..chatroomId(widget.chatroomId)).build()));
-    _addPaginationListener();
+    _chatroomBloc.add(LMChatFetchChatroomEvent(chatroomId: widget.chatroomId));
     scrollController.addListener(() {
       _showScrollToBottomButton();
       _handleChatTopic();
@@ -112,25 +98,6 @@ class _LMChatroomScreenState extends State<LMChatroomScreen> {
       MarkReadChatroomEvent(chatroomId: widget.chatroomId),
     );
     super.dispose();
-  }
-
-  _addPaginationListener() {
-    pagedListController.addPageRequestListener(
-      (pageKey) {
-        _conversationBloc.add(
-          LoadConversations(
-            getConversationRequest: (GetConversationRequestBuilder()
-                  ..chatroomId(widget.chatroomId)
-                  ..page(pageKey)
-                  ..pageSize(500)
-                  ..isLocalDB(false)
-                  ..minTimestamp(0)
-                  ..maxTimestamp(currentTime))
-                .build(),
-          ),
-        );
-      },
-    );
   }
 
   void _handleChatTopic() {
@@ -186,147 +153,6 @@ class _LMChatroomScreenState extends State<LMChatroomScreen> {
     });
   }
 
-  void updatePagingControllers(LMChatConversationState state) {
-    if (state is ConversationLoaded) {
-      _page++;
-
-      if (state.getConversationResponse.conversationMeta != null &&
-          state.getConversationResponse.conversationMeta!.isNotEmpty) {
-        conversationMeta
-            .addAll(state.getConversationResponse.conversationMeta!);
-      }
-
-      if (state.getConversationResponse.conversationAttachmentsMeta != null &&
-          state.getConversationResponse.conversationAttachmentsMeta!
-              .isNotEmpty) {
-        Map<String, List<LMChatMedia>> getConversationAttachmentData = state
-            .getConversationResponse.conversationAttachmentsMeta!
-            .map((key, value) {
-          return MapEntry(
-            key,
-            (value as List<dynamic>?)
-                    ?.map((e) => LMChatMedia.fromJson(e))
-                    .toList() ??
-                [],
-          );
-        });
-        conversationAttachmentsMeta.addAll(getConversationAttachmentData);
-      }
-
-      if (state.getConversationResponse.userMeta != null) {
-        userMeta.addAll(state.getConversationResponse.userMeta!);
-      }
-      List<Conversation>? conversationData =
-          state.getConversationResponse.conversationData;
-      // filterOutStateMessage(conversationData!);
-      conversationData = addTimeStampInConversationList(
-          conversationData, chatroom.communityId!);
-      if (state.getConversationResponse.conversationData == null ||
-          state.getConversationResponse.conversationData!.isEmpty ||
-          state.getConversationResponse.conversationData!.length > 500) {
-        pagedListController.appendLastPage(conversationData ?? []);
-      } else {
-        pagedListController.appendPage(conversationData ?? [], _page);
-      }
-    }
-    if (state is ConversationPosted) {
-      addConversationToPagedList(
-        state.postConversationResponse.conversation!,
-      );
-    } else if (state is LocalConversation) {
-      addLocalConversationToPagedList(state.conversation);
-    } else if (state is ConversationError) {
-      toast(state.message);
-    }
-    if (state is ConversationUpdated) {
-      if (state.response.id != lastConversationId) {
-        addConversationToPagedList(
-          state.response,
-        );
-        lastConversationId = state.response.id;
-      }
-    }
-  }
-
-  // This function adds local conversation to the paging controller
-  // and rebuilds the list to reflect UI changes
-  void addLocalConversationToPagedList(Conversation conversation) {
-    List<Conversation> conversationList =
-        pagedListController.itemList ?? <Conversation>[];
-
-    if (pagedListController.itemList != null &&
-        conversation.replyId != null &&
-        !conversationMeta.containsKey(conversation.replyId.toString())) {
-      Conversation? replyConversation = pagedListController.itemList!
-          .firstWhere((element) =>
-              element.id ==
-              (conversation.replyId ?? conversation.replyConversation));
-      conversationMeta[conversation.replyId.toString()] = replyConversation;
-    }
-    conversationList.insert(0, conversation);
-    if (conversationList.length >= 500) {
-      conversationList.removeLast();
-    }
-    if (!userMeta.containsKey(user!.id)) {
-      userMeta[user!.id] = user;
-    }
-
-    pagedListController.itemList = conversationList;
-    rebuildConversationList.value = !rebuildConversationList.value;
-  }
-
-  void addConversationToPagedList(Conversation conversation) {
-    List<Conversation> conversationList =
-        pagedListController.itemList ?? <Conversation>[];
-
-    int index = conversationList.indexWhere(
-        (element) => element.temporaryId == conversation.temporaryId);
-    if (pagedListController.itemList != null &&
-        conversation.replyId != null &&
-        !conversationMeta.containsKey(conversation.replyId.toString())) {
-      Conversation? replyConversation = pagedListController.itemList!
-          .firstWhere((element) =>
-              element.id ==
-              (conversation.replyId ?? conversation.replyConversation));
-      conversationMeta[conversation.replyId.toString()] = replyConversation;
-    }
-    if (index != -1) {
-      conversationList[index] = conversation;
-    } else if (conversationList.isNotEmpty) {
-      if (conversationList.first.date != conversation.date) {
-        conversationList.insert(
-          0,
-          Conversation(
-            isTimeStamp: true,
-            id: 1,
-            hasFiles: false,
-            attachmentCount: 0,
-            attachmentsUploaded: false,
-            createdEpoch: conversation.createdEpoch,
-            chatroomId: chatroom.id,
-            date: conversation.date,
-            memberId: conversation.memberId,
-            userId: conversation.userId,
-            temporaryId: conversation.temporaryId,
-            answer: conversation.date ?? '',
-            communityId: chatroom.communityId!,
-            createdAt: conversation.createdAt,
-            header: conversation.header,
-          ),
-        );
-      }
-      conversationList.insert(0, conversation);
-      if (conversationList.length >= 500) {
-        conversationList.removeLast();
-      }
-      if (!userMeta.containsKey(user!.id)) {
-        userMeta[user!.id] = user;
-      }
-    }
-    pagedListController.itemList = conversationList;
-    rebuildConversationList.value = !rebuildConversationList.value;
-  }
-
   @override
   Widget build(BuildContext context) {
     ScreenSize.init(context);
@@ -341,20 +167,13 @@ class _LMChatroomScreenState extends State<LMChatroomScreen> {
             bloc: _chatroomBloc,
             listener: (context, state) {
               if (state is LMChatroomLoadedState) {
-                chatroom = state.getChatroomResponse.chatroom!;
-                lastConversationId =
-                    state.getChatroomResponse.lastConversationId ?? 0;
+                chatroom = state.chatroom;
+                lastConversationId = state.lastConversationId;
                 _chatroomActionBloc
                     .add(MarkReadChatroomEvent(chatroomId: chatroom.id));
-                _conversationBloc.add(InitConversations(
-                  chatroomId: chatroom.id,
-                  conversationId: lastConversationId,
-                ));
                 LMAnalytics.get().track(
                   AnalyticsKeys.syncComplete,
-                  {
-                    'sync_complete': true,
-                  },
+                  {'sync_complete': true},
                 );
                 LMAnalytics.get().track(AnalyticsKeys.chatroomOpened, {
                   'chatroom_id': chatroom.id,
@@ -366,9 +185,8 @@ class _LMChatroomScreenState extends State<LMChatroomScreen> {
             },
             builder: (chatroomContext, chatroomState) {
               if (chatroomState is LMChatroomLoadedState) {
-                final response = chatroomState.getChatroomResponse;
-                chatroom = response.chatroom!;
-                actions = response.chatroomActions!;
+                chatroom = chatroomState.chatroom;
+                actions = chatroomState.actions;
                 return Column(
                   children: [
                     widget.appbarbuilder?.call(
@@ -377,126 +195,8 @@ class _LMChatroomScreenState extends State<LMChatroomScreen> {
                         ) ??
                         _defaultAppBar(chatroom),
                     Expanded(
-                      child: BlocConsumer<LMChatConversationBloc,
-                              LMChatConversationState>(
-                          bloc: _conversationBloc,
-                          listener: (context, state) {
-                            updatePagingControllers(state);
-                            if (state is ConversationPosted) {
-                              Map<String, String> userTags =
-                                  LMChatTaggingHelper.decodeString(state
-                                          .postConversationResponse
-                                          .conversation
-                                          ?.answer ??
-                                      "");
-                              LMAnalytics.get().track(
-                                AnalyticsKeys.chatroomResponded,
-                                {
-                                  "chatroom_type": chatroom.type,
-                                  "community_id": chatroom.communityId,
-                                  "chatroom_name": chatroom.header,
-                                  "chatroom_last_conversation_type": state
-                                          .postConversationResponse
-                                          .conversation
-                                          ?.attachments
-                                          ?.first
-                                          .type ??
-                                      "text",
-                                  "tagged_users": userTags.isNotEmpty,
-                                  "count_tagged_users": userTags.length,
-                                  "name_tagged_users": userTags.keys
-                                      .map((e) => e.replaceFirst("@", ""))
-                                      .toList(),
-                                  "is_group_tag": false,
-                                },
-                              );
-                            }
-                            if (state is ConversationError) {
-                              LMAnalytics.get().track(
-                                AnalyticsKeys.messageSendingError,
-                                {
-                                  "chatroom_id": chatroom.id,
-                                  "chatroom_type": chatroom.type,
-                                  "clicked_resend": false,
-                                },
-                              );
-                            }
-                            if (state is MultiMediaConversationError) {
-                              LMAnalytics.get().track(
-                                AnalyticsKeys.attachmentUploadedError,
-                                {
-                                  "chatroom_id": chatroom.id,
-                                  "chatroom_type": chatroom.type,
-                                  "clicked_retry": false
-                                },
-                              );
-                            }
-                            if (state is MultiMediaConversationPosted) {
-                              LMAnalytics.get().track(
-                                AnalyticsKeys.attachmentUploaded,
-                                {
-                                  "chatroom_id": chatroom.id,
-                                  "chatroom_type": chatroom.type,
-                                  "message_id": state.postConversationResponse
-                                      .conversation?.id,
-                                  "type": mapMediaTypeToString(
-                                      state.putMediaResponse.first.mediaType),
-                                },
-                              );
-                            }
-                          },
-                          builder: (context, state) {
-                            return ValueListenableBuilder(
-                              valueListenable: rebuildConversationList,
-                              builder: (context, value, child) {
-                                return PagedListView(
-                                  pagingController: pagedListController,
-                                  scrollController: scrollController,
-                                  physics: const ClampingScrollPhysics(),
-                                  padding: EdgeInsets.symmetric(
-                                    vertical: 10,
-                                    horizontal: 2.w,
-                                  ),
-                                  reverse: true,
-                                  builderDelegate:
-                                      PagedChildBuilderDelegate<Conversation>(
-                                    noItemsFoundIndicatorBuilder: (context) =>
-                                        const Center(
-                                      child: LMChatText('No chats found!'),
-                                    ),
-                                    firstPageProgressIndicatorBuilder:
-                                        (context) =>
-                                            const LMChatSkeletonChatList(),
-                                    animateTransitions: true,
-                                    transitionDuration:
-                                        const Duration(milliseconds: 500),
-                                    itemBuilder: (context, item, index) {
-                                      if (item.isTimeStamp != null &&
-                                              item.isTimeStamp! ||
-                                          item.state != 0 &&
-                                              item.state != null) {
-                                        return _defaultStateBubble(
-                                          item.state == 1
-                                              ? LMChatTaggingHelper
-                                                  .extractFirstDMStateMessage(
-                                                  item.toConversationViewData(),
-                                                  user!.toUserViewData(),
-                                                )
-                                              : LMChatTaggingHelper
-                                                  .extractStateMessage(
-                                                  item.answer,
-                                                ),
-                                        );
-                                      }
-                                      return item.userId == user!.id
-                                          ? _defaultSentChatBubble(item)
-                                          : _defaultReceivedChatBubble(item);
-                                    },
-                                  ),
-                                );
-                              },
-                            );
-                          }),
+                      child:
+                          LMChatConversationList(chatroomId: widget.chatroomId),
                     ),
                     LMChatroomBar(
                       chatroom: chatroom,
@@ -510,32 +210,6 @@ class _LMChatroomScreenState extends State<LMChatroomScreen> {
           ),
         ),
       ),
-    );
-  }
-
-  Widget _defaultStateBubble(String message) {
-    return LMChatStateBubble(message: message);
-  }
-
-  Widget _defaultSentChatBubble(Conversation conversation) {
-    return LMChatBubble(
-      conversation: conversation.toConversationViewData(),
-      currentUser:
-          (LMChatPreferences.instance.getCurrentUser as User).toUserViewData(),
-      conversationUser: conversation.member!.toUserViewData(),
-      onTagTap: (tag) {},
-      isSent: true,
-    );
-  }
-
-  Widget _defaultReceivedChatBubble(Conversation conversation) {
-    return LMChatBubble(
-      conversation: conversation.toConversationViewData(),
-      currentUser:
-          (LMChatPreferences.instance.getCurrentUser as User).toUserViewData(),
-      conversationUser: conversation.member!.toUserViewData(),
-      onTagTap: (tag) {},
-      isSent: false,
     );
   }
