+++ lib/packages/ui/.flutter-plugins
@@ -1,6 +1,4 @@
 # This is a generated file; do not edit or check into version control.
-connectivity_plus=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/connectivity_plus-6.0.3/
-device_info_plus=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/device_info_plus-10.1.0/
 media_kit_video=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/media_kit_video-1.2.4/
 package_info_plus=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/package_info_plus-8.0.0/
 path_provider=/Users/divyanshgandhi/.pub-cache/hosted/pub.dev/path_provider-2.1.3/
